+++ scim2_client/engines/httpx2.py
@@ -96,6 +96,8 @@ class SyncSCIMClient(BaseSyncSCIMClient):
     :param client: A :class:`httpx2.Client` instance that will be used to send requests.
     :param provider: The :class:`~scim2_models.ScimProvider` describing the server.
         If a request payload describe a resource it does not know, an exception will be raised.
+        The :class:`~scim2_models.ScimPolicy` it carries rules how much the payloads
+        exchanged with the server may depart from the specification.
     :param resource_models: Deprecated, pass a :paramref:`provider` instead.
     :param check_request_payload: If :data:`False`,
         :code:`resource` is expected to be a dict that will be passed as-is in the request.
@@ -358,6 +360,8 @@ class AsyncSCIMClient(BaseAsyncSCIMClient):
     :param client: A :class:`httpx2.AsyncClient` instance that will be used to send requests.
     :param provider: The :class:`~scim2_models.ScimProvider` describing the server.
         If a request payload describe a resource it does not know, an exception will be raised.
+        The :class:`~scim2_models.ScimPolicy` it carries rules how much the payloads
+        exchanged with the server may depart from the specification.
     :param resource_models: Deprecated, pass a :paramref:`provider` instead.
     :param check_request_payload: If :data:`False`,
         :code:`resource` is expected to be a dict that will be passed as-is in the request.